Move App component to own directory
